+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Master
 
+- Added official support for Rails 7 and 8
+- Added support for Ruby 3.0 through 3.4
+
 # 0.1.2 (26-8-2020)
 
 - Add build badge to readme